トップレイアウト修正
