+++ app/views/layouts/_header.html.erb
@@ -87,10 +87,15 @@
         </div>
       </nav>
       <%= video_tag("/videos/production ID_3977086.mp4", autoplay: :true, loop: :true, muted: :true) %>
-        <div class="text-sub-center-top owner-btn">
-          <p class="text1">美味しい物が食べたいけど、時間とお金がない！</p>
-          <p class="text2">食事する事で時間と出費が節約出来ます。</p>
+      <div class="text-sub-center-top owner-btn">
+        <div class="row">
+          <div class="col-sm-12">
+            <h1 class="text1">　おいしいお店巡り<br>嬉しいお家外食</h1>
+            <h1 class="text2">手抜きじゃない！！時短で楽してお家でも手軽に</h1><br>
+            <h1 class="text2">厳選されたお店で楽しくおいしい食事を</h1>
+          </div>
         </div>
+      </div>
     </div>
   <% end %>
 </header>
