+++ app/views/private_stores/edit.html.erb
@@ -137,12 +137,6 @@
     <%= form.select :delivery, [['不可', false],['可', true]], { selected: @private_store.delivery }, class: 'form-control col-sm-2' %>
   </div>
 
-  <div class="field">
-    <%= form.label "フードロスにご協力していただけますか？" %><br>
-    <%= form.select :food_loss, [['いいえ', false],['はい', true]], { selected: @private_store.food_loss }, class: 'form-control col-sm-2' %>
-  </div>
-
-
   <div class="actions">
     <%= form.submit %>
   </div>
